Use custom resoure for self signed cert
