+++ pkg/helpers/certificate.go
@@ -102,91 +102,3 @@ func CertificateSelfSigned(host string) (*tls.Certificate, error) {
 
 	return &cert, nil
 }
-
-type TLSPemCertBytes struct {
-	CACert []byte
-	CAKey  []byte
-	Cert   []byte
-	Key    []byte
-}
-
-func SelfGeneratedCertsForDocker(validFor time.Duration) (*TLSPemCertBytes, error) {
-	rsaBits := 2048
-
-	capriv, err := rsa.GenerateKey(rand.Reader, rsaBits)
-	if err != nil {
-		return nil, err
-	}
-
-	notBefore := time.Now().Add(-1 * time.Hour)
-	notAfter := notBefore.Add(validFor)
-
-	ca := &x509.Certificate{
-		SerialNumber: big.NewInt(1),
-		Subject: pkix.Name{
-			CommonName: "Convox",
-		},
-		NotBefore: notBefore,
-		NotAfter:  notAfter,
-
-		IsCA:                  true,
-		KeyUsage:              x509.KeyUsageDigitalSignature | x509.KeyUsageCertSign | x509.KeyUsageKeyEncipherment,
-		BasicConstraintsValid: true,
-	}
-
-	caBytes, err := x509.CreateCertificate(rand.Reader, ca, ca, &capriv.PublicKey, capriv)
-	if err != nil {
-		return nil, err
-	}
-
-	caCertificate, err := x509.ParseCertificate(caBytes)
-	if err != nil {
-		return nil, err
-	}
-
-	caPemBytes := pem.EncodeToMemory(&pem.Block{Type: "CERTIFICATE", Bytes: caBytes})
-
-	cakeyBytes := x509.MarshalPKCS1PrivateKey(capriv)
-	if err != nil {
-		return nil, err
-	}
-	cakeyPemBytes := pem.EncodeToMemory(&pem.Block{Type: "RSA PRIVATE KEY", Bytes: cakeyBytes})
-
-	serverCert := &x509.Certificate{
-		SerialNumber: big.NewInt(2),
-		Subject: pkix.Name{
-			CommonName: "Convox",
-		},
-		NotBefore: notBefore,
-		NotAfter:  notAfter,
-
-		ExtKeyUsage:           []x509.ExtKeyUsage{x509.ExtKeyUsageServerAuth, x509.ExtKeyUsageClientAuth},
-		KeyUsage:              x509.KeyUsageDigitalSignature | x509.KeyUsageKeyEncipherment,
-		BasicConstraintsValid: true,
-	}
-
-	serverpriv, err := rsa.GenerateKey(rand.Reader, rsaBits)
-	if err != nil {
-		return nil, err
-	}
-
-	serverBytes, err := x509.CreateCertificate(rand.Reader, serverCert, caCertificate, &serverpriv.PublicKey, capriv)
-	if err != nil {
-		return nil, err
-	}
-
-	certPemBytes := pem.EncodeToMemory(&pem.Block{Type: "CERTIFICATE", Bytes: serverBytes})
-
-	keyBytes := x509.MarshalPKCS1PrivateKey(serverpriv)
-	if err != nil {
-		return nil, err
-	}
-	keyPemBytes := pem.EncodeToMemory(&pem.Block{Type: "RSA PRIVATE KEY", Bytes: keyBytes})
-
-	return &TLSPemCertBytes{
-		CACert: caPemBytes,
-		CAKey:  cakeyPemBytes,
-		Cert:   certPemBytes,
-		Key:    keyPemBytes,
-	}, nil
-}
