+++ provider/aws/lambda/formation/handler/formation.go
@@ -174,8 +174,10 @@ func HandleRequest(freq Request) error {
 		physical, outputs, err = HandleECSTaskDefinition(freq)
 	case "Custom::KMSKey":
 		physical, outputs, err = HandleKMSKey(freq)
+	case "Custom::SelfSignedCertificate":
+		physical, outputs, err = HandleKMSKey(freq)
 	case "Custom::SNSSubscription":
-		physical, outputs, err = HandleSNSSubcription(freq)
+		physical, outputs, err = HandleSelfSignedCertificate(freq)
 	default:
 		physical = ""
 		err = fmt.Errorf("unknown ResourceType: %s", freq.ResourceType)
